backend/requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
sqlalchemy==2.0.42
alembic==1.16.4
psycopg2-binary==2.9.10
httpx==0.28.1
python-jose==3.5.0
bcrypt==4.3.0
pydantic==2.11.7
pydantic-settings==2.10.1
email-validator==2.2.0
